+++ src/app/app/receipts/[id]/loading.tsx
@@ -1,0 +1,10 @@
+export default function Loading() {
+  return (
+    <div className="space-y-4">
+      <div className="v-skel h-4 w-24" />
+      <div className="v-skel h-[180px] rounded-2xl" />
+      <div className="v-skel h-[260px] rounded-2xl" />
+      <div className="v-skel h-[180px] rounded-2xl" />
+    </div>
+  );
+}